Robot tests: make click in contentmenu actions more specific.
When I try the original in a real browser, I get a click on *all* a-tags within the actions menu.
This means that several popup windows appear.
Clicking only the immediately contained a-tag means only the menu is opened.
I hope this fixes https://github.com/collective/collective.easyform/issues/244