Travis: test on both 5.2.x and 5.2.3-pending.
On those versions, pin check-manifest to 0.44 to prevent getting version conflicts for `build`, `pep517` and `virtualenv`.

We want to test with newer and older zope.interface versions, for https://github.com/collective/collective.easyform/issues/252
5.2.x is currently 5.2.2, with zope.interface 5.0.2.
5.2.3-pending has zope.interface 5.1.2.

This will fail once 5.2.3-pending is out of pending, but that is just a signal that I can clean this Travis config up again.