Removed testModificationsToPropSheetLinesNotPuged.
This test makes no sense in Plone 5.2.
It passed until now because 'qi.reinstallProducts' does nothing since Plone 5.1.

The site properties are not really used anymore in Plone itself, so we could probably remove this from the default profile.
But let's keep it to be safe. This is a maintenance branch.