Use batching on the saved data page.
On a client site I have a form with 12 questions and almost 700 entries and this takes over 40 seconds to load.
Downloading the csv data takes just 4 seconds.

By default the page now uses a batch size of 10.
For good measure, you can override this by passing `?batch_size=x` in the url, but this gets lost when you use the batch navigation to go to the next page.
This is because it is not supported by the standard `CrudBatchView` from `plone.z3cform`.